Removing references to deprecated spring security password encoder
