+++ onebusaway-users/src/main/java/org/onebusaway/users/impl/authentication/VersionedPasswordEncoder.java
@@ -15,42 +15,39 @@
  */
 package org.onebusaway.users.impl.authentication;
 
-import org.springframework.dao.DataAccessException;
-import org.springframework.security.authentication.encoding.PasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 public class VersionedPasswordEncoder implements PasswordEncoder {
 
-  private static final String DEFAULT_SEPARATOR_CHARACTER = "|";
+	private static final String DEFAULT_SEPARATOR_CHARACTER = "|";
 
-  private String _versionId;
+	private String _versionId;
 
-  private PasswordEncoder _passwordEncoder;
+	private PasswordEncoder _passwordEncoder;
 
-  private String _separatorCharacter = DEFAULT_SEPARATOR_CHARACTER;
+	private String _separatorCharacter = DEFAULT_SEPARATOR_CHARACTER;
 
-  public void setVersionId(String versionId) {
-    _versionId = versionId;
-  }
-  
-  public void setPasswordEncoder(PasswordEncoder passwordEncoder) {
-    _passwordEncoder = passwordEncoder;
-  }
+	public void setVersionId(String versionId) {
+		_versionId = versionId;
+	}
 
-  @Override
-  public String encodePassword(String rawPass, Object salt)
-      throws DataAccessException {
-    return _versionId + _separatorCharacter
-        + _passwordEncoder.encodePassword(rawPass, salt);
-  }
+	public void setPasswordEncoder(PasswordEncoder passwordEncoder) {
+		_passwordEncoder = passwordEncoder;
+	}
 
-  @Override
-  public boolean isPasswordValid(String encPass, String rawPass, Object salt)
-      throws DataAccessException {
+	@Override
+	public String encode(CharSequence rawPassword) {
+		return _versionId + _separatorCharacter
+				+ _passwordEncoder.encode(rawPassword);
+	}
 
-    int index = encPass.indexOf(_separatorCharacter);
-    if (index == -1)
-      return false;
-    encPass = encPass.substring(index + 1);
-    return _passwordEncoder.isPasswordValid(encPass, rawPass, salt);
-  }
+	@Override
+	public boolean matches(CharSequence rawPassword, String encodedPassword) {
+		int index = encodedPassword.indexOf(_separatorCharacter);
+		if (index == -1)
+			return false;
+		encodedPassword = encodedPassword.substring(index + 1);
+		return _passwordEncoder.matches(rawPassword, encodedPassword);
+	}
+
 }
