+++ onebusaway-users/src/main/resources/org/onebusaway/users/application-context.xml
@@ -140,25 +140,15 @@
         </property>
     </bean>
     
-    <bean id="passwordEncoderV1" class="org.onebusaway.users.impl.authentication.VersionedPasswordEncoder">
-        <property name="versionId" value="v1" />
-        <property name="passwordEncoder">
-            <bean class="org.springframework.security.authentication.encoding.ShaPasswordEncoder">
-                <constructor-arg value="256" />
-                <property name="encodeHashAsBase64" value="true" />
-            </bean>
-        </property>
-    </bean>
-
-    <bean id="passwordCryptoEncoderV1" class="org.onebusaway.users.impl.authentication.VersionedCryptoPasswordEncoder">
-        <property name="versionId" value="v1" />
+    <bean id="passwordEncoderV2" class="org.onebusaway.users.impl.authentication.VersionedPasswordEncoder">
+        <property name="versionId" value="v2" />
         <property name="passwordEncoder">
             <bean id="encoder" class="org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder"/>
         </property>
     </bean>
 
     <bean id="userService" class="org.onebusaway.users.impl.UserServiceImpl">
-        <property name="passwordEncoder" ref="passwordEncoderV1" />
+        <property name="passwordEncoder" ref="passwordEncoderV2" />
     </bean>
 
     <bean id="userLastAccessTimeServiceCache" class="org.onebusaway.container.spring.ehcache.EhCacheFactoryBean">
